feat(archival): only expose wrap operations
This diff improves the abstraction of the internal/archival package
to only expose wrappers for netx types as opposed to exposing custom
operations. The new code is obviously more abstract because we only
modify well-known types and there's no new operation to learn.

This diff WILL need to be applied to ooni/probe-cli.