refactor(measurex): more understandable model
While in general we may have many lookups, it's fair to say that
in this case we just have a single DNS lookup.

So, warn if we have more than one, and otherwise save the lookup if
we end up with just a single DNS lookup.

Also, write accessors to get internals in a safe way.