fix(websteps): more robust heuristic for IPv6 not being available
I have only seen net_unreach/host_unreach being used for censoring
once. We have said multiple times we'd like to do a better job
in the pipeline for flagging failures. In light of that, it makes
sense to have a more aggressive heuristic here for removing this
source of false positives inside the probe.