fix(websteps): simple heuristic to detect legitimate redirects
If the HTTP response is a redirect and the location and original URL
are equal (plus or minus "www."), the redirect is legitimate.

In such cases, there's no point going down the DiffHTTP rabbit
hole, which may also fail in case the same endpoint behaves
differently for the TH and the probe. (This could possibly be
an indication of a transparent proxy but we should not
definitely end up flagging this case as a DiffHTTP.)