feat: add dnsping implementation
The implementation is flexible enough that we can use it as a
standalone implementation or inside websteps.

I have not added support for the TTL but possibly we can also do
that in the future, should we need a dnstraceroute.

While there, add binary to perform dnsping so we have a way
to easily test this code and experiment.

While there, rename QUICListener to UDPListener because this is
definitely a better name. Do not rename the definition of the
"quic_listen" event, our of an abundance of caution.

While there, reset the measurex DNS timeout to 4 seconds
because we now have DNS ping as a backup :-).