fix(websteps): truly enforce max number of addrs per family
They key to do that is always returning a stable list with the
addresses initially discovered by the probe first.

We'll amend this list with HTTP3 info, if this is needed.

Then, we'll append to this list new IP addresses previously not
known to the probe discovered using dnsping/TH.

While there, ensure we can choose a more aggressive profile in
terms of speed (--fast) or precision (--deep).

Also, a previous commit in c/w/main.go completely broke the
semantics of --deep, but we're also taking care of it now by
basically making the wrong --deep become --fast and by
reintroducing the values previously used for --deep.