feat: use non-connected socket for UDP resolver
This diff modifies the UDP resolver inside netxlite to use a
non-connected socket. We are doing this because we would like
to see a better error than "connection_refused" when there's
some ICMP error telling us (probably) that the target port
is unreachable.

While working on this topic, ensure there's not a lot of code
duplication by doing the following:

1. factoring out code to ParseUDPAddr that was used both by
internal/netxlite/quic.go and internal/dnsping;

2. ensuring we use the same UDP resolver engine to implement both
the UDP resolver and the dnsping.

While there, discover that we were not validating the query ID
of received replies :facepalm: and ensure now we do that.

While there, recognize that we could use a better name for all the
structs implementing a given kind of DNS transport.

While there, recognize that we should definitely skip the dnsping
entries that failed when building the URLAddress lists.