fix(websteps): implement stable redirect chains
If we're measuring http://torproject.org with the default config we'll
end up using just two of the four A and two of the four AAAA addrs.

The next step redirects to https://torproject.org, which is bound to the
same four A and four AAAA addresses.

In such a case, we want the second step to reuse the same addresses we
used in the first step, to construct realistic-looking redirects.

This design of course simplifies web-connectivity-like data analysis.