feat: implement NS lookup resolution
At the measurex level, you can ask for a NS lookup by using
a flag that extends the performed queries.

At the archival and netxlite levels, we implement a new function
that matches the prototype provided by the standrad library.

We're not using the stdlib, though, because it's not clear yet
in which platforms a NS lookup is reliably possible.

(IIRC, the stdlib uses `/etc/resolv.conf` in many Unix cases
including Android and iOS, where it won't work.)

Of course, we also needed to introduce support for decoding NS
queries, which is easy thanks to miekg/dns.

We're implementing NS resolutions because we aim to use it
in the TH to inform the probe about NS servers.

The probe will use this information to spot parked domains (we'll
have a list of known parking NSs and filter using that).

Because parked domains are false positives in many cases, this is
quite useful even though we'll probably just know some of them.