refactor: use a transport also for system resolver
The advantage of using a (obviously fake) transport for the system
resolver is that we can collect its results more easily.

In fact, we can now tap into any resolution at the transport level
and we can more easily emit the OONI data format for lookups.

The previous implementation of the system resolver has been moved
to i/n/legacy.go because legacy code still needs it.