refactor: emit DNS round trips using OONI data format
This diff modifies the emitted data format so that:

1. we only emit DNS round trips

2. we emit them using the OONI data format

This change is made possible by the previous diff, which ensured
we were able to treat all resolvers uniformly.

In particular, now we see the system resolver as a resolver using
its custom transport, so we can collect query (and query-like)
data directly at the transport level.