refactor(measurex): introduce AbstractMeasurer
The AbstractMeasurer is an interface that replaces a concrete
Measurer with an abstract version of the Measurer concept.

This functionality is obviously great for testing. Now, in fact,
websteps does not depend on a concrete Measurer anymore. And
the same holds for the dnsping package.

However, I am not introducing this new functionality for testing
but rather because it seems possible to use AbstractMeasurer
to implement caching.

In turn, I probably need caching for two reasons:

1. to speed up reprocessing of measurements locally when I
introduce new algorithms for flagging results;

2. to build up knowledge about all the URLs in the test list,
which is likewise useful to better reason about our common
input set and apply specific algorithms.