feat: implement {dns,endpoint} measurement caching
We import the codebase used by Go to maintain its package cache, which
has been modified to remove the `log.txt` logfile.

We removed `log.txt` because it quickly grows quite big.

We define rules for equality of dns plans and dns measurements as
well as endpoint plans and endpoint measurements.

It turns out a proper rule for endpoint measurements is a bit tricky
because of options, so we exclude such endpoints from caching.

We wrap a measurer with a new measurer that is cache aware.

The cache aware measurer will consult the cache before trying to
measure an endpoint or perform a dns lookup.

After a measurement, the measurer will also update the cache.

We also relaxed cookie rules. Basically, now we only check the names
of the cookies rather than their content.

This relaxing actually could probably help to measure a bit less.

I am doing this for two reasons:

1. add a caching layer with expiring rules to the TH

2. add a caching later w/o any expiring rule to the probe so that
I can trace once and then repeat the measurement multiple times

The latter seems quite useful if the objective is that of tuning the
algorithms inside of measurex/analysis.go.