refactor(measurex): use single DNSLookupPlan per LookupType
We need to refactor the code until there is a single plan for
each measurement. This is the best and only way to ensure that
we can easily map a plan to its results.

While there, since I have been a bit stuck, revamp the
dnslookup command to ensure I understand what I see
in output and refactor the code to add documentation
and ensure the code is safe and we have all the
abstractions we actually need.