feat: change DNS change policy
We're now caching on a domain basis. We have improved the rules
to check whether a plan and a measurement are related as well as
the rules for comparing two DNS measurements.

Note that the latter rules are stronger than the ones we use
when matching measurements. Here we require the two measurements
to be using the same resolver. We cannot obviously do the same
for comparing probe and TH since the TH always uses DoH.

While there, notice that there are multiple definitions of network
and ensure we only use the definition inside archival.