refactor: change endpoint cache policy
Store all endpoints by IP address and check for equality using
the current summary (we still need to do some work to allow for
weird endpoints with SNI/Host/ALPN to slip in).

While there, make the dns implementation more similar to the
endpoint implementation by introducing a summary for it.