feat: extract CNAME from TH responses
This change requires adding support for extracting the CNAME from
a DNS round trip result. We store the discovered CNAME into the DNS
lookup result, which is what the probe uses. (We leave the real
round trip almost unparsed and we only gather the fields we need
in the more flat DNS lookup result.)

While there, notice that THRequestAsync requires a buffered channel
to properly handle context, so use a buffered channel.

While there, write some syntactic sugar for interacting with the
THClient in a much nicer and simpler way.

While there, restructure thctl to also emit the archival data format.

While there, flag that now we expect to use the DNS round trip to
generate the archival data format. Generating the archival data format
from the DNS round trip looks much more natural. We have now use for
serializing the DNS lookup result to the archival data format, except
guaranteeing backwards compatibility to legacy code.

While there, recognize the we can implement THResponseWithID as a
THResponse with a non-JSON-serializable URLMeasurementID.

While there, ensure that a DNS reply is a reply when we parse it.