fix(measurex): remove caching endpoints restrictions
Create a summary of an endpoint that also includes all the
relevant options. So, we're now also matching options.

Add extra logging around this functionality because I predict
there may be cases in which it does not work, so having logging
should generally be enough to sort out any issue.