fix: ensure that caching works as intended
We cannot exclude the possibility that a dns measurement and an
endpoint measurement could have a has collision.

If this happens, the cache does not work as intended.

Hence, split the cache in a dns cache and an endpoint cache.

Additionally, the serialization of alpn depends on options and may
either be `null` or `[]`. This fact is obviously bad in terms of
caching, so enforce a consistent representation but always returning
an empty array whenever the ALPN is not set.