fix(websteps): limit max number of address tested by the TH
We should not exaggerate with the number of IP addresses tested by
the test helper. Let's add one more IPv4 and one more IPv6 to the
probe's knowledge, to cover the case where the probe is blocked
or has discovered only bogons or blockpages, but let's not measure
dozens of IP addresses per domain. Doing that would frankly be
too aggressive on the test helper's behalf.