fix(measurex): use correct referer and headers
The correct referer is the previous URL rather than the location.

We should not pass extra headers because the TH will actually
reject most headers, hence matching will fail.

(Regarding the latter point, I can probably make the TH reject a
request with wrong URLs, to additionally be on the safe side?)