refactor: replace logger with logcat singleton
This commit replaces model.Logger instances pretty much everywhere with
calls to the logcat package, which contains a singleton.

A nice property of the logcat is that it buffers quite an amount of
messages. For now, it is not as nice as Android's logcat, though, but
I've named it as such because I think that's the final objective.

This way of logging would be quite useful in mobile apps, where we
are currently struggling to deliver the logs in all cases and we
especially fail when it matters the most (early bootstrap).

Anyway, the real reason for making a logcat in this repository is that
I need more log levels to more nicely separate websteps normal output
from explanations from more nerdy output.

The Debug / Info / Warn levels of apex/log just are not enough.

So, with the new logcat we have enough levels now.

Also, we're not throwing away the existing model.Logger, rather we
are gently making it less important. It will eventually be there
to be the destination where the logcat sends logs. It matters having
a model.Logger for that, because model.Logger implements the same
interface used by apex/logger, which we use in the probe.