fix(netxlite): avoid embedding for resolver
With embedding, every time we add a new method, we don't get a compiler
error, which means the new method is pass through.

In most cases, we want to modify or perform actions for methods, thus it's
actually better not to have this form of automatic wrapping.

Noticed because I could not realize why LookupPTR returned an unwrapped
error in the TH. It turns out this was the root cause.

While there, try to improve the situation for other types inside of netxlite,
without trying to be super exhaustive. I just fixes all the types that I could
easily spot after a single pass over the whole of netxlite.