refactor(atomicx): make it a ~zero-cost abstraction
We can reduce the complexity of Int64 by just ensuring its int64 value is
the first element: this is enough to guarantee alignment.

This allows us to drop the mutex and really reduce the cost of calling int64
atomic functions. It is not really zero cost but the cost is less.

Also, this solution is more elegant.