feat(websteps): allow filling fully reusable measurements cache
This diff introduces changes allowing one to use the websteps command
to fill a predictable, complete cache from a given location.

The cache is predictable if you force predictable resolvers, otherwise
every run may use a different resolver.

The cache you obtain in this way is also complete. This means that
it contains both the TH's and the probe's measurements.

In turn, the availability of these information, allow a person to
spin up a local `thd` and use a local `websteps` command to basically
re-run the original measurement using the cache.

This functionality is great in terms of re-running known test
cases and testing variations of the scoring algorithms.

Here's what changed in detail:

1. allow `cmd/websteps` to specify a probe cache and a TH cache

2. allow `cmd/websteps` to force predictable resolvers

3. in `websteps/client.go` allow someone to change the resolvers
after creating a new `Client`

4. in `websteps/client.go` allow someone to configure a hook
to observe TH results

5. in `measurex/cache.go` expose more functionality from the `Cache`
so one can save the TH results from the `Client` hook