feat: implement dnsping caching
We _also_ need to cache dnsping results to make measurements fully
reproducible later in time. Without this functionality, websteps is
going to attempt dnspings on its own skewing the results.

While there, because we're putting more results into the cache, let's
ise better names for directories containing dns and endpoint
results: "dns" and "endpoint" respectively are more clear than "d" and "e.

While there, notice that the predictable resolver function was not
setting _also_ the system resolver, so modify it to set such a resolver.