fix(measurex): make NewURLAddressList deterministic
This matters when we're rerunning previous measurements using the cache.

Without this fix, the TH will randomly pick IP addresses to test, and
this means that in a bunch of cases the probe coild not match its
measurements against the TH measurements.

While the TH always includes the probe's IP addresses into its own
measurements, it will also select _other_ IP addresses to test.

We do that so that the probe will have additional addresses to test
just in case all of the probe's resolvers are lying.

Yet, if the extra IP addresses we pick are not deterministic with
respect to the originally cached DNS resolution, the TH, which will
run without network capabilities, will see cache misses when the
ordering of the IP addresses choosen by that particular run doesn't
match the order of the cached run.

This happens because we're using a map in NewURLAddressList to
ensure we count each IP address once and we flag all that we already
know for such an IP address (e.g., HTTP3 support).

This patch fixes NewURLAddressList to use a custom ordered map
so that we NewURLAddressList always emits IP addresses in the
same order in which they appeared.

In turn, the order in which they appear depends on DNS lookups,
and the TH does just a single lookup using DoH.

When we're using the cache, the single DoH lookup does not
actually occur but is the one we cached previously.

Thus, the order used by NewURLAddressList matches the order
of IP addresses inside the cached lookup.