fix(p/t/shell): make rerun work again
My previous rewrite broke the way in which we set content into
the cache. We need to serialize a list of values.

We were trying to serialize objects, so JSON failed.

Even fixing the issue and trying to serialize raw objects (i.e., the
original dicts from which we created objects) is wrong.

In fact, cache entries are lists not dicts.

So, the simplest solution is to store the raw content of the cache
at the cache level and allow iterating over that.

After applying this fix, I rerun some experiments and the result
looks good (i.e., same as not using the cache), so I think we are
fine now as long as this rerun failure bug was concerned.