fix(websteps): more pragmatic approach to flags
Keep the low 32 bits for anomalies. Keep up high 32 bits for things
the probe noted and may be useful but are not anomalies.

It certainly helps to take decisions. In fact we now have a function that
tells us when we have an anomaly and we use it.

It will also certainly help the pipeline.

Of course, also update our definition of flags in Python code.