fix: ensure we write all the logs
When we're using a slow mechanism to transfer logs (e.g., ssh) we may
end the main program much before we've seen al logs using the shell.

As a fix, introduce synchronization between the main and the log
consumer(s), such that we wait for then to finish.

Tested by running tests over SSH with a slow connection to the
machine I was connected to using SSH.