fix(measurex): return cached results sorted by completion time
I've seen cases where a resolver returns A, B, C and another one
returns C, B, A where the first resolver completes first.

If the probe picks up the second from the cache, it is going to
have cache misses because the probe using the cache will try using
C and B while the original probe used A and B.

To fix, record when a DNS or endpoint measurement completes and
sort the results ascending when accessing the cache.

This means a probe repeating a measurement should see the results
from the cache in the same order of the original probe.

Let's now test this diff.