fix(python): avoid pylance type errors
Yesterday, pylance was able to infer that a List[DBLikeEntry] is
a List[Tabulable] because DBLikeEntry.tabular() exists and has the
correct signature (i.e., returns Tabular).

Today, this is not true anymore.

In the meanwhile, VSCode and its plugins update.

I don't know whether this is a regression or an improvement
for pylance, but in any case, we can rewrite the code so that
we don't need to typecast an entire list.

To this end, we just need to define a new function that does
the conversion element by element inside dblike.py and we
update all call sites to avoid using mapcreate.