feat(dbsteps): ci/cu: include measurements samples
This diff automates part of my workflow for going through
measurements and checking whether the results are consistent
or we have any false positive or negative.

I used to list all interesting tags for a single measurement
using `ci`, take note of each tag, then search for tag to
obtain a sample of the measurements.

Finally, I searched each of them.

All of this happened on paper and it was quite slow.

With this diff, I can get a sample of the measurements to
check for each tag _directly_.

I think I'll commit more enhancements to simplify my workflow
soon, but this functionality is already nice!