feat(dbsteps): implement interactive ci or cu
This command adds an interactive prompt allowing us to cycle
through each row returned by ci or cu.

The interactive form of ci is cii.

The interactive form of cu is cui.

The cii and cui commands basically stop after each row and give
is a prompt to further investigate the sampled measurements.

Such a prompt accepts a limited version of the commands available
in the main environment, which should be sufficient to look into
the current row's measurements and find false positives or negatives.

Because we only sample four measurement per category, it should be
reasonably possible for a human to remember which of the four
have already been inspected manually.

This diff attempts to automate another part of my workflow for
inspecting measurements for false positives and negatives.