fix: clarify the terms of comparison
Since it's a 1:1 comparison, it does not make much sense to
say "medium" because I do not have other benchmarks.

It makes instead sense to say what I think would be the
webconnectivity behavior compared to websteps.

(TBH, I should also run measurements to understand whether
this is the case, but for now having these notes there helps
me to know which aspects to keep in mind.)